feat: complete P1 diagnostics and resource-fit profile

+++ sdk/examples/timer1_encoder_demo/main.c
@@ -14,7 +14,7 @@ int main(void) {
 
   if (!omcu_hw_abi_is_compatible(OMCU_HW_ABI_MAJOR) ||
       !omcu_hw_has_feature(required) ||
-      !omcu_tn9k_timer1_configure(0u, UINT32_MAX, 4u, ctrl)) {
+      !omcu_tn9k_timer1_configure(0u, UINT16_MAX, 4u, ctrl)) {
     for (;;) {
     }
   }
